Player info added to the UI in consolidated format.
