+++ Assets/Scripts/Management/UIStates/UIGameState.cs
@@ -6,12 +6,12 @@ public class UIGameState : UIBaseState
 {
     public override void EnterState(UIManager context)
     {
-        
+        context.GameUI.SetActive(true);
     }
 
     public override void ExitState(UIManager context)
     {
-        
+        context.GameUI.SetActive(false);
     }
 
     public override BaseState<UIManager> UpdateState(UIManager context)
