--- Assets/Scripts/UI/UIWeaponInfoController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b7a3f09e8cd628d4dbacef4459f62e2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
